% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break-package.R
\docType{package}
\name{break-package}
\alias{break}
\alias{break-package}
\title{break: Interactive Watershed Break Point Delineation}
\description{
Interactive Shiny app for clicking stream networks to place break points, delineating upstream watersheds, and exporting sub-basins via pairwise subtraction. Built on the fresh package for FWA-referenced spatial hydrology queries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NewGraphEnvironment/break}
  \item \url{https://newgraphenvironment.github.io/break/}
  \item Report bugs at \url{https://github.com/NewGraphEnvironment/break/issues}
}

}
\author{
\strong{Maintainer}: Al Irvine \email{al@newgraphenvironment.com}

Other contributors:
\itemize{
  \item New Graph Environment Ltd. [copyright holder]
}

}
\keyword{internal}
